# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OFFITEC ERP API",
    "version": "1.0.0",
    "description": "OFFITEC Kurumsal ERP Sistemi API Dokümantasyonu"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Hata mesajı burada yazar."
          }
        }
      },
...
      "CreateEmployeeRequest": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ],
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Ahmet"
          },
          "lastName": {
            "type": "string",
            "example": "Yilmaz"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "ahmet.yilmaz@offitec.com"
          },
          "password": {
            "type": "string",
            "example": "SecurePassword123!"
          },
          "title": {
            "type": "string",
            "example": "Senior Developer"
          },
          "departmentId": {
            "type": "string",
            "example": "it-dept"
          },
          "roleName": {
            "type": "string",
            "example": "Mühendis"
          },
          "phone": {
            "type": "string",
            "example": "+90 555 123 4567"
          },
          "address": {
            "type": "string",
            "example": "İstanbul, Türkiye"
          },
          "hireDate": {
            "type": "string",
            "format": "date-time",
            "example": "2026-01-15T00:00:00Z"
          },
          "annualLeaveEntitlement": {
            "type": "integer",
            "example": 14
          },
          "notes": {
            "type": "string",
            "example": "Personel hakkında not"
          }
        }
      },
      "UpdateEmployeeRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "departmentId": {
            "type": "string"
          },
          "roleName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "annualLeaveEntitlement": {
            "type": "integer"
          },
          "notes": {
            "type": "string"
          },
          "terminationDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/employees": {
      "post": {
        "tags": [
          "Employees"
        ],
        "summary": "Yeni çalışan oluştur",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Çalışan oluşturuldu"
          },
          "400": {
            "description": "Geçersiz veri",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Yetkisiz"
          },
          "403": {
            "description": "Erişim reddedildi"
          }
        }
      },
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Çalışan listesini getir",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            },
            "description": "Aktif / pasif filtresi"
          },
          {
            "in": "query",
            "name": "departmentId",
            "schema": {
              "type": "string"
            },
            "description": "Departman ID"
          },
          {
            "in": "query",
            "name": "roleName",
            "schema": {
              "type": "string"
            },
            "description": "Rol filtresi"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Ad/soyad/e-posta arama"
          }
        ],
        "responses": {
          "200": {
            "description": "Çalışan listesi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Yetkisiz"
          },
          "403": {
            "description": "Erişim reddedildi"
          }
        }
      }
    },
    "/employees/{id}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Personel detayını getir",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Personel ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Personel detayı"
          },
          "404": {
            "description": "Personel bulunamadı"
          }
        }
      },
      "patch": {
        "tags": [
          "Employees"
        ],
        "summary": "Personel bilgilerini güncelle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Personel ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Personel güncellendi"
          },
          "400": {
            "description": "Geçersiz veri"
          },
          "404": {
            "description": "Personel bulunamadı"
          }
        }
      }
    },
...
  },
...
}